+++ engine/test/rgbscript/rgbscript_test.cpp
@@ -365,7 +365,7 @@ void RGBScript_Test::runScripts()
                 // Unknown, new and RGBScriptProperty::None are not valid
                 QVERIFY(property.m_type == RGBScriptProperty::List ||
                         property.m_type == RGBScriptProperty::Range ||
-                        property.m_type == RGBScriptProperty::Integer ||
+                        property.m_type == RGBScriptProperty::Float ||
                         property.m_type == RGBScriptProperty::String);
                 // Check property specificities
                 switch (property.m_type)
@@ -445,7 +445,7 @@ void RGBScript_Test::runScripts()
                         }
                     }
                     break;
-                case RGBScriptProperty::Integer:
+                case RGBScriptProperty::Float:
                     // Test with an integer value
                     s.setProperty(property.m_name, QString::number(-1024));
                     qDebug() << "  Readback: " << s.property(property.m_name);
